+++ app/src/test/java/comnickdchee/github/a3am/ExampleUnitTest.java
@@ -17,39 +17,4 @@ public class ExampleUnitTest {
     public void addition_isCorrect() {
         assertEquals(4, 2 + 2);
     }
-    Book b = new Book("1234567891234", "HarryPotter", "J R R martin");
-
-    @Test
-    public void ISBN_correct() {
-        {
-            assertEquals(b.getISBN().length(), 13);
-        }
-    }
-    @Test
-    public void getAuthorName() {
-        {
-            assertEquals(b.getAuthor(), "J R R martin");
-        }
-    }
-    @Test
-    public void getTitle() {
-        {
-            assertEquals(b.getTitle(), "HarryPotter");
-        }
-    }
-    @Test
-    public void setAuthorName() {
-        {
-            b.setAuthor("J K Rowlin");
-            assertEquals("J K Rowlin", b.getAuthor());
-        }
-    }
-    @Test
-    public void setTitle() {
-        {
-            b.setTitle("Harry");
-            assertEquals("Harry", b.getTitle());
-        }
-    }
-
 }
